+++ CHANGELOG.md
@@ -146,6 +146,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - "/api/agent" endpoint. #2542
 - "/api/exploitations/manual" endpoint. #2509
 - "/api/island/ip-addresses" endpoint. #2565
+- ElasticSearch fingerprinter. #2674
 
 ### Fixed
 - A bug in network map page that caused delay of telemetry log loading. #1545